1.0.0

+++ cj.netos.absorb.robot.program/src/cj/netos/absorb/robot/IHubTailBillService.java
@@ -2,6 +2,7 @@ package cj.netos.absorb.robot;
 
 import cj.netos.absorb.robot.model.HubTails;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IHubTailBillService {
@@ -11,11 +12,11 @@ public interface IHubTailBillService {
 
     List<HubTails> pageBillOfMonth(String wenyBankID, int order, int year, int month, int limit, long offset);
 
-    long totalInBillOfMonth(String wenyBankID, int year, int month);
+    BigDecimal totalInBillOfMonth(String wenyBankID, int year, int month);
 
-    long totalInBillOfYear(String wenyBankID, int year);
+    BigDecimal totalInBillOfYear(String wenyBankID, int year);
 
-    long totalOutBillOfMonth(String wenyBankID, int year, int month);
+    BigDecimal totalOutBillOfMonth(String wenyBankID, int year, int month);
 
-    long totalOutBillOfYear(String wenyBankID, int year);
+    BigDecimal totalOutBillOfYear(String wenyBankID, int year);
 }
